feat(analytics): fold overview into the hub and retire the advanced route
PR0b of the analytics redesign. The standalone Analytics overview page
becomes the Overview tab of the registry-driven hub, the retired
/ui/analytics/advanced route redirects into the hub, and the sidebar
collapses to a single Analytics entry. Frontend only; no API changes.

- New components/AnalyticsOverview.svelte holds the former overview
  (summary StatCards, top-species / time-of-day / trend / new-species
  charts, recent-detections table), now driven by the hub's shared date
  range instead of its own FilterForm.
- pages/Analytics.svelte is the hub (moved from the deleted
  AdvancedAnalytics.svelte); it renders AnalyticsOverview for the Overview
  tab and defaults to Overview, so /ui/analytics stays query-clean.
- Species auto-select is gated on the active tab filtering by species, so
  the Overview tab never writes a species URL param.
- navigation.svelte.ts gains redirect() (a replaceState variant of
  navigate). App.svelte redirects /ui/analytics/advanced to
  /ui/analytics?tab=patterns, preserving existing query params.
- DesktopSidebar.svelte drops the redundant Advanced Analytics nav item.
- Removes the now-unused pageTitle.advancedAnalytics and
  analytics.advanced.title i18n keys across all locales.

The Overview tab now shares the hub date-range presets
(week/month/quarter/year/custom, default month) instead of the old page's
own filter, so the former all-time and today presets are gone and the
default shifts from week to month. Stale ?range= values fall back to the
month default.

Tests: hub tests updated (default Overview tab, coming-soon now on Review
& Accuracy, species auto-select gating); the #978 stale-response race
test moves to AnalyticsOverview and retriggers via a date-range change;
navigation redirect() unit test added.
