+++ frontend/src/lib/desktop/features/analytics/components/__tests__/StubOverview.svelte
@@ -1,0 +1,3 @@
+<!-- Test fixture: a trivial stand-in for AnalyticsOverview so hub tests can
+     assert the Overview tab mounts without pulling in its data fetching / D3. -->
+<div data-testid="stub-overview">overview panel</div>
