+++ frontend/static/messages/pl.json
@@ -205,7 +205,6 @@
     "serverError": "500 - Wewnętrzny błąd serwera",
     "componentError": "Błąd ładowania komponentu",
     "speciesAnalytics": "Analityka gatunków",
-    "advancedAnalytics": "Zaawansowana analityka",
     "detectionDetails": "Szczegóły wykrycia",
     "settingsNotAvailable": "Ustawienia niedostępne"
   },
@@ -1786,7 +1785,6 @@
       }
     },
     "advanced": {
-      "title": "Zaawansowana analityka",
       "chartControls": "Kontrolki wykresu",
       "dateRange": "Zakres dat",
       "chartOptions": "Opcje wykresu",
